Add copyright blurbs to sources
We've had a COPYING file since 78335c57 (COPYING: Add the 2-clause BSD
license, 2014-09-21), but these blurbs ensure the license stays
attached to the content even if it ends up being separated from this
repository.  If the injected bookmarklet preserved newlines, that
would make reading the license easier, but at least it's there ;).